refactor definition tests into separate subsuite
