+++ test/files/scalacheck/quasiquotes/Test.scala
@@ -9,4 +9,6 @@ object Test extends Properties("quasiquotes") {
   include(PatternDeconstructionProps)
   include(LiftableProps)
   include(ErrorProps)
+  include(DefinitionConstructionProps)
+  include(DefinitionDeconstructionProps)
 }
